% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitparseR.R
\name{fitparseR}
\alias{fitparseR}
\title{fitparseR: A package for reading garmin .fit files}
\description{
The fitparseR package uses reticulate and jsonlite to wrap
  the python fitparse package, which seems to be kept more
  current than the native R solutions.
}
\section{read .fit file}{

  \link{get_fit_dfs}
}

