% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_dfs.R
\name{get_fit_dfs}
\alias{get_fit_dfs}
\title{read gps fit data files}
\usage{
get_fit_dfs(fitfilename, checkconda = TRUE)
}
\arguments{
\item{fitfilename}{name (including file path) of .fit file}

\item{checkconda}{logical, check conda environment for packages}
}
\value{
a list of three dataframes:  \eqn{session},
    \eqn{records}and \eqn{events}
}
\description{
\code{get_fit_dfs} reads a gps .fit files using python
 and put into session, data record, and event dataframes
}
